typo

+++ asafov.andrew/P4/countUniqLetters.cpp
@@ -4,12 +4,12 @@ int lab::countUniqLetters(const char* string, size_t length)
 {
   char* newstring = new char[length];
   int count = 0;
-  for (int i = 0; i < length; i++)
+  for (size_t i = 0; i < length; i++)
   {
     if((string[i]>=64 && string[i]<=91)||(string[i]>=97 && string[i]<=122))
     {
       int a = 0;
-      for(int j =0; j<length; j++)
+      for(size_t j =0; j < length; j++)
       {
         if (string[i] == newstring[j])
         {
